Extract helper functions
This commit extracts helper function to reduce the scope of the Command
subclasses in the Keys namespace down to just passing config and
arguments. This separation of concerns is great because it allows me to
mock these helper methods and test the commands in isolation and then
also test the helper functions with DI so that I can simulate different
situations.
