+++ src/commands/keys/add.ts
@@ -1,6 +1,6 @@
 import { Command } from "@oclif/command"
 import { Jay } from "../../shared/Jay"
-import { computeSafelist } from "../../helpers/keys"
+import { updateSafelist } from "../../helpers/keys"
 
 export default class Add extends Command {
   static description = "add a user to the safelist (then update)"
@@ -8,20 +8,7 @@ export default class Add extends Command {
   static strict = false
 
   async run(): Promise<void> {
-    const { safelistFilePath } = Jay.config
-
     const appendList = this.parse(Add).argv
-
-    const existingData = Jay.utils.readFile(safelistFilePath)
-    const initialList = existingData.split("\n")
-
-    const safeList = computeSafelist({
-      appendList,
-      initialList,
-      removeList: [],
-    })
-    const data = safeList.join("\n")
-
-    Jay.utils.writeFile(safelistFilePath, data)
+    updateSafelist(Jay.config, appendList, [])
   }
 }
