+++ src/commands/keys/init.ts
@@ -1,28 +1,11 @@
 import { Command } from "@oclif/command"
 import { Jay } from "../../shared/Jay"
+import { initKeys } from "../../helpers/keys"
 
 export default class Init extends Command {
   static description = "make directory structure and symlink"
 
   async run(): Promise<void> {
-    const {
-      keysDir,
-      keysFilePath,
-      safelistFilePath,
-      symlinkPath,
-      unmanagedFilePath,
-    } = Jay.config
-
-    Jay.utils.mkdir(keysDir)
-
-    const paths: string[] = [safelistFilePath, keysFilePath, unmanagedFilePath]
-
-    paths.forEach((path) => {
-      Jay.utils.writeFile(path, "")
-    })
-
-    if (!Jay.utils.fileExists(symlinkPath)) {
-      Jay.utils.symlink(keysFilePath, symlinkPath)
-    }
+    initKeys(Jay.config)
   }
 }
