+++ src/commands/keys/raw.ts
@@ -1,5 +1,6 @@
 import { Command } from "@oclif/command"
 import { Jay } from "../../shared/Jay"
+import { addRawKey } from "../../helpers/keys"
 
 export default class Raw extends Command {
   static description = "add a raw key to the unmanaged list (then update)"
@@ -7,16 +8,7 @@ export default class Raw extends Command {
   static strict = false
 
   async run(): Promise<void> {
-    const { unmanagedFilePath } = Jay.config
-
     const rawPath: string = this.parse(Raw).argv[0]
-    this.log(rawPath)
-
-    let unmanagedKeys = Jay.utils.readFile(unmanagedFilePath)
-    const newRawKey = Jay.utils.readFile(rawPath)
-
-    unmanagedKeys = [unmanagedKeys, newRawKey].join("\n")
-
-    Jay.utils.writeFile(unmanagedFilePath, unmanagedKeys)
+    addRawKey(Jay.config, rawPath)
   }
 }
